Allows admins to skip some paper sanitization (#72553)
This PR allows admins with R_FUN to create paper that doesn't sanitize
as much HTML as the default, player-accessible paper. Specifically, the
ability for admins with R_FUN to add images that papercode would
normally sanitise via HTML img tags.

I'd like to make some fancy papers to send as faxes, but the current
sanitization doesn't allow for images.

![image](https://user-images.githubusercontent.com/41448081/211176186-c33d611d-8ac3-4683-bef8-10016e34eaf4.png)

This was a fax I threw together in a few minutes after making the logo
:cl:
admins: Admins with the appropriate permissions can now use HTML image
tags in paper and faxes.
/:cl: