Fix photocopying removing stamp overlays from the original instead of copying them to the copy (#88764)
Photocopiers were trying to copy the stamp overlays from the new blank
paper onto the original, and cutting the original's old overlays in the
process.
This inverts that statement and removes the cutting part.

Nice if photocopying doesn't remove the stamp overlays on the original
just cause it feels like it.
:cl:
fix: Photocopying no longer removes the stamp overlays from the original
paper, and actually copies them to the copy.
/:cl: