Changes the jobs Tzimisce can do (#259)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

I found it pretty silly that tzimisce could play a role like sheriff and
prince but could not play a more basic role like hound, I decided to fix
that, tzim can now play hound, keeping being able to play sheriff but
removes their ability to play prince and seneschal
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More inline with lore and made at the suggestion of the vision team
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->
<!-- You can uncomment line 1 @ _maps/_basemap.dm to boot up a test map
that loads much faster. -->
<details>
<summary>Screenshots&Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.
Spawning as hound and sheriff

![image](https://github.com/user-attachments/assets/cbeee657-63c9-4adc-8d62-e81d35dcef63)
View from job picker

![image](https://github.com/user-attachments/assets/dede8091-0169-4d43-8d5f-f3a589f64729)

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:

add: Added the ability for Tzimisce to play hound
del: Removed the ability for Tzimisce to play seneschal and prince


<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

Co-authored-by: WaifuCommand <201161694+WaifuCommand@users.noreply.github.com>