makes broken tzimisce floor uncraftable (BUGFIX) (#265)
I swear I tested it



## About The Pull Request
comments out the tzimisce crafting recipe for the necropolis floor
## Why It's Good For The Game

Upon being crafted the tzimisce floor would be created but the
"Crafting..." text would still show on the crafting menu along with the
turning gear icon. This will be replaced by a structure in a future
update.

## Testing Photographs and Procedure

No testing image provided because it is just commenting out a single
line.

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>

## Changelog

Make the tzimisce necropolis floor uncraftable 
del: Removed old things

/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->