update debug server to v2.3.5 (#301)
Fixes the debugger on 516. I'm not sure how old ours is, but it's not
working as intended.

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Updates the debugger to
https://github.com/willox/auxtools/releases/tag/v2.3.5, which fixes a
bunch of stuff on BYOND 516. I'm not even sure how old our current one
is, but it's certainly not working 100% as intended.

No player facing changes.

## Why It's Good For The Game

testing good.

## Evidence
<details>


![VSCodium_DqCdwJBIzy](https://github.com/user-attachments/assets/56b77bf4-ef03-47ab-a429-5a2ca2d24175)

</details>