TheaterExpandedRemap (#317)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

**Theater Remap & JazzClub Fixes & Prep Work to Separate the two
buildings**

New updated style for the theater with much more space for audience.
Fixing many of the complaints and adding some suggestions I've heard
about the JazzClub/Ventrue Domain.
For example, added the new bathroom on the bottom floor, and fixing the
upstairs bar to be functional.
Shoutout to Bane of Totems | Ken North on discord for his help and work
on the jazz club Bar.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
With the new found server pop, I thought it was a good idea to go ahead
and rush along the Theater idea and remap I had in mind, so it could
support more people in the theater, and get feedback for improvements. I
think this will add a bit more to the Theater/Kindred Trial experience.


<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Testing Photographs and Procedure
![Screenshot 2025-04-17
221051](https://github.com/user-attachments/assets/9418ec2f-a918-4e55-854b-06e96e41c640)
![Screenshot 2025-04-17
221059](https://github.com/user-attachments/assets/8df9cfb8-4f32-4ff1-943e-5fd0423aa4c6)
![Screenshot 2025-04-17
221824](https://github.com/user-attachments/assets/5706fc45-68b5-46b8-bcaa-e86bab5ce931)
![Screenshot 2025-04-17
221913](https://github.com/user-attachments/assets/154d31f2-aa5a-4503-8940-946471183692)
![Screenshot 2025-04-17
221926](https://github.com/user-attachments/assets/3bcab97e-4c76-45e8-a073-c055277b5f9c)

## Changelog

SanFransicsco.dmm

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and its effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->